Move audit code into one package
This makes it easier to work on audit features as tests can be run as a set
and it's simple to see all the classes involved in audit logging.

Also moved AuditEventSaveTastTest into Robolectric from Espresso.
