+++ collect_app/src/main/java/org/odk/collect/android/utilities/DialogUtils.java
@@ -29,7 +29,7 @@ import android.widget.ListView;
 import org.odk.collect.android.R;
 
 import org.odk.collect.android.application.Collect;
-import org.odk.collect.android.logic.AuditEvent;
+import org.odk.collect.android.formentry.audit.AuditEvent;
 import org.odk.collect.android.logic.FormController;
 import timber.log.Timber;
