+++ collect_app/src/test/java/org/odk/collect/android/formentry/audit/AuditConfigTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.odk.collect.android.logic;
+package org.odk.collect.android.formentry.audit;
 
 import org.junit.Test;
